fix(ci): python coverage shards — clear stale merge attention
